+++ CHANGELOG.md
@@ -4,16 +4,22 @@ This format is based on [Keep A Changelog](https://keepachangelog.com/en/1.0.0).
 
 ## 3.11.0 -- 2022-10-11
 
+### Added
+
+* `FixedDecimal` Haskell equivalent to `PFixedDecimal`, along with `Num` and
+* `Fractional` instances, `fixedNumerator` and `fixedDenominator` extractor functions.
+
 ### Modified
 
 * `PFixedDecimal` is updated so that it represents decimal point range in exponential form.
 
   Following type and functions are added along with Plutarch numerical instances.
   - `PFixedDecimal`
+  - `pfixedNumerator`
+  - `pfixedDenominator`
   - `pconvertExp`
   - `pfromFixedDecimal`
   - `ptoFixedDecimal`
-  - `pfromRational`
   - `ptoRational`
   - `punsafeMkFixedDecimal`